requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.10.0
click==8.1.3
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==5.0.0
Markdown==3.4.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.2
platformdirs==2.5.4
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
whitenoise==6.2.0
zipp==3.10.0
